+++ test_app/__init__.py
@@ -1,0 +1,6 @@
+"""
+Test app for django-document-manager
+
+This app contains test models and is only used during testing.
+It should NOT be included in the package distribution.
+"""